Fix clean regex typo
The problem is with the hyphen between _ and \u3000. I'm sure it's not intentional, because this way the next hyphen between \u3000 and \u303f is taken as a literal hyphen, not a range.
I think escaping the hyphen after the underline and adding {} explicitly (since it was implicitly allowed so until now) should be fine. But characters between _ and \u3000 were allowed, now these would be disallowed.